[2 - 4단계 방탈출 결제 / 배포] 상돌(이상진) 미션 제출합니다. (#147)
* refactor: Authorization 헤더를 기본 헤더로 설정 및 Read-timeout 시간 수정

* feat: 2단계 클라이언트 코드 추가

* refactor: 특정 날짜와 테마에 대한 모든 시간을 조회하는 API Endpoint 수정 및 처리 메서드 이동

- ReservationController 가 아닌 ReservationTimeController에서 처리하도록 수정
- endpoint를 /reservations가 아닌 /times로 수정

* feat: 예약 조건을 조회하기 위한 별도의 Specification 객체 생성

* refactor: 조건에 맞는 예약을 조회하는 메서드 수정

- 조건의 날짜에 대한 검증 추가

* refactor: 예약과 대기 메서드를 별도로 분리 및 예외 처리 추가

- 예약 대기 API Endpoint를 /reservations/waiting으로 수정
- 예약 요청시 이미 예약이 존재하는지 확인한다.
- 예약 대기 요청시 이미 예약(대기)를 했는지 확인한다.

* refactor: 예약 승인 및 거절 기능 분리

* refactor: 모든 final 키워드 제거

* refactor: Transactional 적용

* refactor: 대기중인 예약을 취소하는 기능을 회원 / 관리자 별로 구현

* test: 추가된 예약 대기 승인 / 취소 / 거절에 대한 테스트 추가

* feat: MemberService에 Transactional 적용

* feat: PaymentClient 인터페이스로 결제 클라이언트를 추상화

* feat: 결제 정보를 저장하는 Entity 구현

* feat: 초기 데이터에 payment 정보 추가

* refactor: 결제 승인 응답 DTO에 승인 날짜 정보 추가

* feat: 결제 취소에 대한 요청 / 응답 DTO 구현

* feat: 결제 취소 응답 파싱에 사용되는 역직렬화 객체 구현

* feat: Payment에 대한 Repository 클래스 구현

* feat: 회원이 자신의 예약을 조회할 때 사용되는 응답 DTO 수정 및 클라이언트 코드 반영

* feat: Payment에 대한 Service 구현

- 결제 정보를 저장하고, 예약 ID로 결제 정보를 삭제한다.

* feat: PaymentClient에 결제 취소 기능 구현

* feat; ReservationRequest에 PaymentRequest를 얻어내는 메서드 추가

* feat: 예약과 Payment 정보를 저장하는 응답 DTO 추가

* feat: ErrorType에 결제 관련 상수 추가

* refactor: 결제 승인 이후 DB 저장 과정에서 예외가 발생하면 결제를 취소하도록 수정

* refactor: ReservationService에서 Payment 기능 분리

- 순수 예약 및 대기만 담당하도록 수정

* feat: 취소된 결제 정보를 저장하는 별도의 Entity 및 Repository 생성

* feat: PaymentKey를 이용하여 Payment 정보를 불러오는 기능 추가

* feat: PaymentService에 결제 취소 및 취소 시간 업데이트 메서드 추가

* feat: 예약과 결제의 트랜잭션 처리를 위한 별도의 Service 객체 생성

- 이를 이용한 컨트롤러 코드 수정

* refactor: 테스트용 데이터의 결제 승인 시간을 현 시간으로 수정

* refactor: 어드민 예약 조회시 확정된 예약만 나오도록 수정 및 일부 Service 메서드 간소화

* refactor: Payment 관련 클래스 패키지 세분화

* refactor: Payment에서 Reservation 필드를 not null로 지정

* feat: 관리자 페이지에 Waiting 바로가기 추가

* feat: 관리자가 대기중인 예약을 승인하면, 해당 예약의 상태를 '확정 - 결제 필요'로 지정

- 이에 따른 유저 예약 페이지 수정

* fix: 유저의 예약 페이지 조회와 예약 승인이 정상적으로 작동하지 않는 문제 해결

* refactor: 결제되지 않은 예약은 결제 API 요청 없이 삭제하도록 수정

* feat: 어드민 예약 페이지에 결제 여부 컬럼 추가

* refactor: 요청 DTO에서의 결제 금액 타입을 String -> Long 으로 수정

* style: 유저의 예약을 조회하는 메서드명 수정

* feat: DTO의 LocalDate, LocalTime 직렬화, 역직렬화 기능 구현 및 이에 따른 JsonFormat 제거

* refactor: 결제의 날짜 타입을 OffsetDateTime으로 변경

* refactor: 예약 결제 이후 에러 발생시 결제 취소 정보가 저장되지 않는 문제 해결

* refactor: 미사용 생성자 제거

* comment: 결제 처리 관련 주석 추가

* refactor: 테스트용 초기 데이터 수정 및 구조 변화로 인해 실패하던 테스트 수정

* refactor: 예약 / 대기 조회 권한 수정

* feat: ReservationController 에서의  예약 추가 / 삭제와 결제 상황에 대한 테스트 추가

* test: PaymentService 테스트 추가

* test: 날짜와 시간 직렬화 / 역직렬화 테스트 추가

* test: Payment 객체 생성시의 예외 검증 확인

* test: CanceledPayment 객체 생성시의 예외 검증 확인

* test: ReservationWithPaymentService에 대한 테스트 추가

* test: 결제 API의 취소 정보 역직렬화 테스트 추가

* refactor: 로그아웃 시 쿠키 값을 null로 바꾸도록 수정 및 테스트

* refactor: Reservation에서의 미사용 메서드 제거 및 예외 검증 로직 수정

* refactor: ReservationSearchSpecification에서의 예약 상태 조회 로직 수정

- Reservation 객체를 생성할 때 status에 대한 null 체크를 하도록 수정한 것을 반영
- 상태값을 입력받아 비교하는 것이 아닌, 확정(결제 대기) / 대기로 조회

* remove: 미사용 클래스 제거

* fix: 관리자가 조건을 이용하여 예약을 조회하는 기능이 작동하지 않는 문제 해결

* test: 관리자가 특정 조건으로 예약을 조회하는 테스트 추가

* feat: 테마를 추가할 때 같은 이름의 테마가 존재하는지 확인

* remove: 미사용 클래스 제거

* refactor: 미사용 메서드 제거

* test: 관리자 페이지 접속 테스트 추가

* test: 회원 페이지 접속 테스트 추가

* test: ThemeService에 대한 테스트 추가

* test: Member 객체 생성자에서의 예외 테스트 추가

* test: PaymentService에서의 예외 테스트 추가

* test: ReservationTimeService에서의 예외 테스트 추가

* test: 예약 대기 추가 / 승인 API 테스트 추가

* test: 결제 클라이언트 테스트 추가

* style: 미사용 import문 제거

* fix: 회원 예약 목록이 나오지 않는 문제 해결

* feat: 예약 대기를 추가할 때 사용하는 별도의 DTO 생성

* refactor: 리프레쉬 토큰 기능 삭제

* rename: ApiResponse의 클래스명 수정(-> RoomEscapeApiResponse)

* feat: Swagger 초기 환경 설정

* refactor: Swagger Schema 설정

* feat: 예약 및 대기 API 문서 작성을 위한 코드 추가

* feat: 인증 / 인가 API 문서 작성을 위한 코드 추가

* feat: 로그인 상태 확인을 위한 인터셉터와 어노테이션 구현

* refactor:  MethodArgumentNotValidException 에 대한 예외 처리 수정

* feat: LocalDate, LocalTime 직렬화 / 역직렬화를 위한 ObjectMapper 빈 등록

* fix: 데이터 롤백이 되지 않아 실패하는 테스트 수정

* refactor: 로그아웃 시 로그인 여부를 확인하도록 수정 및 테스트 추가

* feat: 회원 API 문서 작성을 위한 코드 추가

* refactor: 로그인 / 관리자 interceptor에서 예외 발생시 로그인 페이지로 리다이렉트 하도록 수정

* refactor: 회원 예약 조회 및 예약 페이지 접근시의 로그인 체크 추가

* refactor: 토큰 만료 및 쿠키가 존재하지 않을 때의 메시지 수정

* feat: 예약이 존재하는 테마 삭제시 예외 발생 기능 추가

* feat: 테마 API 문서 작성을 위한 코드 추가

* feat: 예약 시간 API 문서 작성을 위한 코드 추가

* refactor: 관리자 API와 회원 API 태그 추가

* test: 페이지 접근 테스트 추가

* refactor: AuthController 테스트에서의 미사용 테스트 제거

* fix: 관리자 예약 페이지에서의 기능 오작동 해결

- 관리자가 수동으로 예약하면 결제 대기 상태로 예약한다.

* refactor: 날짜 파라미터 입력시 조건 추가

* refactor: Swagger description 수정

* fix: 시간 설정 및 권한 수정으로 인해 실패하는 테스트 수정

* test: 관리자 예약 페이지에서의 예약에 대한 테스트 추가 및 실패하는 테스트 수정

* refactor: 주간 인기 테마 API Endpoint 수정

* style: Swagger Description 내용 수정

* docs: ERD 사진 업로드

* style: Swagger Tag description 변경

* refactor: 회원 예약 조회 페이지 경로 수정

* refactor: 모든 엔티티 클래스의 생성자 접근제어자 수정

* refactor: 메서드 주석 제거

* refactor: 주간 인기 테마 API 엔드포인트 수정

* refactor: API Description 제거

* test: TossPaymentClient의 테스트 수정

* feat: PaymentProperties에 timeout 값 추가

* refactor: PaymentClient 인터페이스 제거

* refactor: TossPaymentClient 테스트 방식 수정(-> RestClientTest 사용)

+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -4,6 +4,8 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import roomescape.reservation.domain.repository.ReservationRepository;
 import roomescape.system.exception.ErrorType;
 import roomescape.system.exception.RoomEscapeException;
 import roomescape.theme.domain.Theme;
@@ -13,19 +15,25 @@ import roomescape.theme.dto.ThemeResponse;
 import roomescape.theme.dto.ThemesResponse;
 
 @Service
+@Transactional
 public class ThemeService {
-    private final ThemeRepository themeRepository;
 
-    public ThemeService(final ThemeRepository themeRepository) {
+    private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
+
+    public ThemeService(ThemeRepository themeRepository, ReservationRepository reservationRepository) {
         this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
-    public Theme findThemeById(final Long id) {
+    @Transactional(readOnly = true)
+    public Theme findThemeById(Long id) {
         return themeRepository.findById(id)
                 .orElseThrow(() -> new RoomEscapeException(ErrorType.THEME_NOT_FOUND,
                         String.format("[themeId: %d]", id), HttpStatus.BAD_REQUEST));
     }
 
+    @Transactional(readOnly = true)
     public ThemesResponse findAllThemes() {
         List<ThemeResponse> response = themeRepository.findAll()
                 .stream()
@@ -35,13 +43,14 @@ public class ThemeService {
         return new ThemesResponse(response);
     }
 
-    public ThemesResponse getTop10Themes(final LocalDate today) {
+    @Transactional(readOnly = true)
+    public ThemesResponse getMostReservedThemesByCount(int count) {
+        LocalDate today = LocalDate.now();
         LocalDate startDate = today.minusDays(7);
         LocalDate endDate = today.minusDays(1);
-        int limit = 10;
 
         List<ThemeResponse> response = themeRepository.findTopNThemeBetweenStartDateAndEndDate(startDate, endDate,
-                        limit)
+                        count)
                 .stream()
                 .map(ThemeResponse::from)
                 .toList();
@@ -49,13 +58,25 @@ public class ThemeService {
         return new ThemesResponse(response);
     }
 
-    public ThemeResponse addTheme(final ThemeRequest request) {
+    public ThemeResponse addTheme(ThemeRequest request) {
+        validateIsSameThemeNameExist(request.name());
         Theme theme = themeRepository.save(new Theme(request.name(), request.description(), request.thumbnail()));
 
         return ThemeResponse.from(theme);
     }
 
-    public void removeThemeById(final Long id) {
+    private void validateIsSameThemeNameExist(String name) {
+        if (themeRepository.existsByName(name)) {
+            throw new RoomEscapeException(ErrorType.THEME_DUPLICATED,
+                    String.format("[name: %s]", name), HttpStatus.CONFLICT);
+        }
+    }
+
+    public void removeThemeById(Long id) {
+        if (themeRepository.isReservedTheme(id)) {
+            throw new RoomEscapeException(ErrorType.THEME_IS_USED_CONFLICT,
+                    String.format("[themeId: %d]", id), HttpStatus.CONFLICT);
+        }
         themeRepository.deleteById(id);
     }
 }
